Refactor `ascend_config_node` from inner function to standalone function
